drivers:mmc:cadence: Calculate delay-element
commit 055da8a7787603c070a376bd4e6c1dbb04337238 from
git@git.assembla.com:cavium/WindRiver.linux.git

The Cadence SDHCI driver uses a delay element count when the DLL
does not lock.  This patch allows the driver to automatically
calculate the number of delay elements to use to meet the timing
rather than use hard-coded values.  This allows for variation between
chips.  The delay element is not used when the DLL is locked.

Change-Id: I7d8c27612831bf981709ce702061f7924fe5cae7
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/124215
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>