Calculate delay-element
commit 6df9262dfe4fcfbb63f9300f32ab780cf9a6b9de from
git@git.assembla.com:cavium/WindRiver.linux.git

With this patch, the delay element count is calculated rather than
relying on the device tree.  The problem with the device tree is that
this value can vary from chip to chip so by measuring the value to use
the timing values should be more accurate.  Delay elements are only used
when there is no DLL lock, otherwise the DLL is used.

Change-Id: I7d8c27612831bf981709ce702061f7924fe5cae7
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/124157
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>