arm64: zynqmp: Disable Tri-state for SDIO
commit cbc7855f94f4e9e2bad9902279c1d28c229aff0f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Since the zynqmp pinctrl driver now includes support for the
tri-state registers, ensure that the pins needing output-enable
are correctly configured for SOMs.

Currently, there is an issue with the detection of the MMC for
the SOM kv260, resulting in the following error:
ZynqMP> mmc dev 1
Card did not respond to voltage select! : -110

To address this problem, configure the SDIO pins for output-enable
to enable MMC detection.

Signed-off-by: Tejas Bhumkar <tejas.arvind.bhumkar@amd.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>