v4l: xilinx: dprx: Fix incorrect reconfiguration of registers
commit 3dc0d604b2385c7ac7138ed7d968f5f4f5be0488 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

The current driver always skips the initial configuration of HDCP1X IP
Registers because of wrong comparison of 'is_enabled' flag.

The default value of 'is_enabled' flag is always '0' and it will be '1'
only when the driver enables configuration registers. So, the driver
should avoid reconfiguration of registers when 'is_enabled' is '1'.

Signed-off-by: Lakshmi Prasanna Eachuri <lakshmi.prasanna.eachuri@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>