fpga: Fix the reset handling
commit 8f631c5d1f66849028fa8b009cf1bc60b7cc17e9 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

The current reset sequence in the FPGA subsystem first deasserts the reset,
performs AFI configurations, and then asserts the reset before handing
control to individual driver probes. This sequence is not ideal as the
individual driver probes should be called with the reset deasserted. A
pulse is triggered for reset after the afi configurations.

The issue was hidden as we had clock_ignore_unused kernel command line
parameter and in that case for the first time the clock is kept enabled and
the xilfpga does the reset.

Fixes: 909a76fe3d09 ("fpga: Add ps-pl reset handling through xilinx afi driver")
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>