clk: clocking-wizard: calculate dividers fractional parts
commit fa5c05a786967d2a2dc735ac73d2f7d96ca09aba from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Calculate dividers fractional parts to optimally modulate output frequency.
Clocking wizard supports having multiplier m and divisors d and o.
Currently we are not utilising the fractional parts of m and o.
For the pixel clock usecases we need the accuracy to be more.
When there is only one output we can modify m d and o. Now we are adding
support for m and o to have fractional values.

Signed-off-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
Tested-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>