net: axienet: Correct tx_buff size when DRE is not enabled
commit 3518c49bc755b2b613a036d31029bfae10f9456f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

If DRE (Data realignment engine) is not enabled in the DMA h/w,
software has to take care of the alignment of the buffers by
allocating a pool of tx buffers in the driver.
tx_buff contains virtual address of driver allocated buffer for
each descriptor when DRE is disabled.
Remove redundant macro XAE_TX_BUFFERS and use TX_BD_NUM_DEFAULT
to correct size of tx_buff same as number of descriptors.

Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>