drm: xlnx: hdmi: Fix TMDS clock calculation for RGB YUV422 YUV444 formats
commit fc263b3517be41050377f028cc2ae8a53446988f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

In case of 10bpc and above RGB, YUV422 and YUV444 formats are not working
due to the wrong condition check.

As per the HDMI 1.4 specification section 6.5, the color depth for
YUV422 is always 12 bpc, so  pixels are carried at a rate of one
pixel for TMDS clock. TMDS clock calculation for RGB and YUV444
are done based on color depth.

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>