iio: adc: s32cc_adc: fix syncing of dma region
commit 3562bc046bb1dbb5ff7104bc29ae37693af21e9e from
https://github.com/nxp-auto-linux/linux

When using CONFIG_DMA_API_DEBUG, the kernel throws warning about the
ownership of DMA memory.
This happens because the syncing of the dma region is done by the ADC
driver. This is incorect because that region is allocated to the dma
driver.

Use the dma device instead of the ADC device.

Issue: ALB-11315
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>