drivers: mailbox: handle SGI for shared IPI
commit 460f1d4deee175b69db9fc9fc07a27174f9705dd from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

At least one IPI is used in TF-A for communication with PMC firmware.
If this IPI needs to be used by other agents such as RPU then, IPI
system interrupt can't be generated in mailbox driver. In such case
TF-A generates SGI to mailbox driver for IPI notification.

Signed-off-by: Tanmay Shah <tanmay.shah@amd.com>
Signed-off-by: Saeed Nowshadi <saeed.nowshadi@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>