dts: overlay: ov5647: Specify clock-noncontinuous on CSI endpoint
commit 6d523c00412b5c6bc2e3020bbc1b48abd8b68804 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The OV5647 declares that it wants a non-continuous clock in its endpoint
configuration. However, it does not expose this information via
get_mbus_config. Even if it would implement get_mbus_config the
bcm2835-unicam driver would not evaluate the flags field of it.

By also specifying clock-noncontinuous on the CSI endpoint itself, the
CSI driver properly picks up the requests and enables the non-continuous
clock mode.

Signed-off-by: Christian Speich <christian@spei.ch>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>