media: chips-media: wave5: Convert to platform remove callback returning void
commit 54525cbcd7b94a4bdf646ee712f2df65b0fc405d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

[ Upstream commit ad9be163c5ba80e3e923be2b31c70d2c6fa27138 ]

The .remove() callback for a platform driver returns an int which makes
many driver authors wrongly assume it's possible to do error handling by
returning an error code. However the value returned is ignored (apart
from emitting a warning) and this typically results in resource leaks.

To improve here there is a quest to make the remove callback return
void. In the first step of this quest all drivers are converted to
.remove_new(), which already returns void. Eventually after all drivers
are converted, .remove_new() will be renamed to .remove().

Trivially convert this driver from always returning zero in the remove
callback to the void returning variant.

Signed-off-by: Uwe Kleine-KÃ¶nig <u.kleine-koenig@pengutronix.de>
Reviewed-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Signed-off-by: Sebastian Fricke <sebastian.fricke@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>