rpmsg: char: Update local endpt address for virtio-rpmsg backend
commit 6e3d3b577e2db60c865b7f8af518c789317efcd1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The rpmsg char driver creates a local end-point when the actual
endpt device is opened. The virtio-rpmsg backend can dynamically allocate
the local end-point address if the endpt creation is done using the
address RPMSG_ADDR_ANY. This is not reflected in the sysfs src file, so
update the stored address with the allocated address in such a case. This
allows the userspace to be able to retrieve the local end-point
address through sysfs.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>