spi: spi-microchip-core: Fix the number of chip selects supported
commit f63965a302b3eeb654358582daac579ad913dc9b from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

The SPI "hard" controller in PolarFire SoC has eight CS lines, but only
one CS line is wired. When the 'num-cs' property is not specified in
the device tree, the driver defaults to the MAX_CS value, which has
been fixed to 1 to match the hardware configuration; however, when the
'num-cs' property is explicitly defined in the device tree, it
overrides the default value.

Fixes: 9ac8d17694b6 ("spi: add support for microchip fpga spi controllers")
Signed-off-by: Prajna Rajendra Kumar <prajna.rajendrakumar@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>