mmc: sdhci_am654: Fix ITAPDLY for HS400 timing
commit faaa3236c86111a46e5e874e4220f2d49ee50839 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit d3182932bb070e7518411fd165e023f82afd7d25 upstream.

While STRB is currently used for DATA and CRC responses, the CMD
responses from the device to the host still require ITAPDLY for
HS400 timing.

Currently what is stored for HS400 is the ITAPDLY from High Speed
mode which is incorrect. The ITAPDLY for HS400 speed mode should
be the same as ITAPDLY as HS200 timing after tuning is executed.
Add the functionality to save ITAPDLY from HS200 tuning and save
as HS400 ITAPDLY.

Fixes: a161c45f2979 ("mmc: sdhci_am654: Enable DLL only for some speed modes")
Acked-by: Adrian Hunter <adrian.hunter@intel.com>
Link: https://lore.kernel.org/r/20240320223837.959900-8-jm@ti.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Judith Mendez <jm@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>