dt-bindings: gpu: Rename img,powervr to img,powervr-rogue
commit 4fd19920b3d201053fe82dc09ea77abde64c8d06 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

[ Upstream commit 35df039b26ac771a019d2f48abd44020eb2615e2 ]

This version is slightly different from upstream as it removes
references to files not backported for this branch.

This binding will be used for GPUs starting from Series6 (Rogue)
and later. A different binding document will describe Series5.
With that the name "img,powervr" is too generic, rename to
"img,powervr-rogue" to avoid confusion.

Suggested-by: Maxime Ripard <mripard@kernel.org>
Signed-off-by: Andrew Davis <afd@ti.com>
Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Reviewed-by: Frank Binns <frank.binns@imgtec.com>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Message-ID: <20240109171950.31010-2-afd@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>