octeontx2-af: Modify SMQ flush sequence to drop packets
commit 194bb01a22cbe75c81f4005010bc8070bc46660c from
git@git.assembla.com:cavium/WindRiver.linux.git

The current implementation of SMQ flush sequence waits for the packets
in the TM pipeline to be transmitted out of the link. This sequence
doesn't succeed in HW when there is any issue with link such as lack of
link credits, link down or any other traffic that is fully occupying the
link bandwidth (QoS). This patch modifies the SMQ flush sequence in HW to
drop the packets after TL1 level (SQM) instead of polling for the packets
to be sent out of RPM/CGX link. Also did some cleanup for storing TM
schq's in smq_flush_ctx.

Fixes: 5d9b976d4480 ("octeontx2-af: Support fixed transmit scheduler
                      topology")
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Change-Id: Ic8711177faf5916e3d07c6918e9b94ca055a6f14
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/128090
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>