input: edt-ft5x06: Correct prefix length in snprintf
commit fa68c2d4748a48bead9344f7cccc119d7caa068e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

[commit e4631fd5c3c3c3710636dfdf9657bdb188bec51d rpi-linux]

snprintf takes the length of the array that we can print into,
and has to fit the NULL terminator in there too.
Printing the prefix is generally "12-3456 " which is 8 desired
characters (the length of EDT_NAME_PREFIX_LEN) and the NULL.
The space is therefore being truncated to fit the NULL in.

Increase the length snprintf is allowed to use.

Patch Link:
https://github.com/raspberrypi/linux/commit/e4631fd5c3c3c3710636dfdf9657bdb188bec51d

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>