drm/bridge: cdns-dsi: Wait for Clk and Data Lanes to be ready
commit de0415e9312c22d906a22715a1b62f9c97c17431 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Once the DSI Link and DSI Phy are initialized, the code needs to wait
for Clk and Data Lanes to be ready, before continuing configuration.
This is in accordance with the DSI Start-up procedure, found in the
Technical Reference Manual of Texas Instrument's J721E SoC[0] which
houses this DSI TX controller.

If the previous bridge (or crtc/encoder) are configured pre-maturely,
the input signal FIFO gets corrupt. This introduces a color-shift on the
display.

Allow the driver to wait for the clk and data lanes to get ready during
DSI enable.

Fixes: e19233955d9e ("drm/bridge: Add Cadence DSI driver")
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>