regulator: rpi_panel_v2: Add delay on I2C reads
commit f1a2a9a675899c166a75c6e398af8aa5445e5e5f from
https://github.com/raspberrypi/linux.git rpi-6.6.y

As with many microcontrollers, the panel wants to use clock
stretching during a read so that the appropriate return value
can be generated and programmed into the hardware.
With Pi0-3 really not supporting clock stretching, this resulted
in the panel firmware dying.

Insert a delay between the write and the read to give the firmware
a chance to generate the relevant return value.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>