tty: serial: fsl: add separate write_atomic callback
commit 6f82b9f0b1972df3d84fe6d07467a4fc577835c2 from
https://github.com/nxp-auto-linux/linux

The previous write_callback acuqires a spinlock which on full PREEMPT_RT
is changed into an rt_mutex. Therefore, it results in an invalid
possible wait during an atomic context.

Add a separate callback which uses the console_lock_atomic function
provided for this case. Also, don't call uart_write_wakeup because it
also takes a spinlock. Most probably, it also isn't necessary to call it
for the write callback since it relates to console outputting but I will
keep it to preserve the original implementation.

Issue: ALB-11307

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
[fix: use spin_lock/unlock_*() functions to replace
uart_port_lock/unlock_*() functions and drop the write_atomic settings
to compatible with standard kernel]
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>