DTS: bcm2712: don't assume L1 sub-state support at chip level
commit e99d1768b77e77182d5e7ac9c335c7d8cbd88963 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

L1 sub-states require the connection of the CLKREQ# auxiliary signal
which is a PCB-level decision, and also depends on the vagrancies of
adapters/interposers/slots that may exist between the chip and the
endpoint.

If this parameter is present, and the RC driver sees an L1 substate
capability in the endpoint, then the result is a broken link if CLKREQ#
isn't end-to-end connected - regardless of the state of the L1.x enable
bits in the RC.

Board-specific methods (HAT+ overlay, board DTB) should define this
property if required.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>