riscv: dts: microchip: define only one default dma region
commit d8720665214634c1a9ab4a0b869eb0e82fe76730 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

During boot the dma core emits a warning when the default DMA is
redefined:
	Reserved memory: created DMA memory pool at 0x00000000c4000000, size 64 MiB
	OF: reserved mem: initialized node non-cached-low-buffer, compatible id shared-dma-pool
	OF: reserved mem: 0x00000000c4000000..0x00000000c7ffffff (65536 KiB) nomap non-reusable non-cached-low-buffer

	Reserved memory: region for default DMA coherent area is redefined

	Reserved memory: created DMA memory pool at 0x0000001412000000, size 256 MiB
	OF: reserved mem: initialized node non-cached-high-buffer, compatible id shared-dma-pool
	OF: reserved mem: 0x0000001412000000..0x0000001421ffffff (262144 KiB) nomap non-reusable non-cached-high-buffer
This used to be a pr_warn() but is now a WARN(), which will cause some
systems to panic. The last region marked as default that is created will
be marked as default, which in our case is the region at 0x1412000000.
Remove the linux,dma-default property from the 0xc4000000 region as it
has always been ignored.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>