firmware: ti_sci: Unconditionally register reset handler
commit e9c08bfdf21e5d3e9c9a4f71e2b5ff4d4a7c762a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

There was once a limitation that there could only be one system
reset handler. Due to that we only would register this handler
when a non-standard device tree property was found, else we left
the default handler in place (usually PSCI). Now that we can
have multiple handlers, and TI-SCI reset is always available
in the firmware, register this handler unconditionally.

This priority is left at the default so higher priority handlers
(like PSCI) are still attempted first.

Signed-off-by: Andrew Davis <afd@ti.com>
Reviewed-by: Markus Schneider-Pargmann <msp@baylibre.com>
Link: https://lore.kernel.org/r/20240326223730.54639-3-afd@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>