dt-bindings: can: m_can: Add wakeup-source property
commit dfcf442879f7ce3f3c24be12dd1150f4b763b42d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

m_can can be a wakeup source on some devices. Especially on some of the
am62* SoCs pins, connected to m_can in the mcu, can be used to wakeup
the SoC.

This property defines on which devices m_can can be used for wakeup.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>