dt-bindings: serial: 8250_omap: Add wakeup pinctrl state
commit 426ed3f5b3b21fbfee391953c3e9cf3d9dc4831f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Pins associated with the 8250 omap unit can be the source of a wakeup in
deep sleep states. To be able to wakeup, these pins have to be
configured in a special way. To support this configuration add the
default and wakeup pinctrl states.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>