drm/tidss: Add dispc_is_idle()
commit 8e76a266feee641ade27877f476a0c6266668aa3 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a helper function, dispc_is_idle(), which returns whether the DSS is
idle (i.e. is any video port enabled).

For now we add a call to it in the suspend and resume callbacks, and
print a warning if in either place the DSS is not idle.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>