drm: bridge: cdns-mhdp8546-core: Move mode_valid hook to bridge_funcs
commit d8a7a0c2ad3363d5e13c769bc18a9ae4777889e3 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

With the support for the 'DRM_BRIDGE_ATTACH_NO_CONNECTOR' case,
the connector_helper funcs are not initialized if the encoder has this
flag in its bridge_attach call. Till now we had mode_valid hook only in
the drm_connector_helper_funcs. Move this hook in drm_bridge_funcs to
validate the modes in both case based on the flag.

Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>