arm64: boot: dts: ti: k3-*: Add memory node to bootloader stage
commit 26e670c5d29f40931827a9a5b76b834687a389a1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the bootph-all property to the memory node so that it can be
accessed by FDT functions at bootloader stage.

The bootloader requires the memory node to be able to initialize and set
the size of the DRAM banks. For this purpose, make sure all memory nodes
are present and standardized, and modify them if not.

Signed-off-by: Neha Malcom Francis <n-francis@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>