ASoC: ti: davinci-mcasp: Set min period size using FIFO config
commit f08ebde41b3053f02de99e994a2b49dfcb494085 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The minimum period size was enforced to 64 as older devices integrating
McASP with EDMA used an internal FIFO of 64 samples.

With UDMA based platforms this internal McASP FIFO is optional, as the
DMA engine internally does some buffering which is already accounted for
when registering the platform. So we should read the actual FIFO
configuration (txnumevt/rxnumevt) instead of hardcoding frames.min to
64.

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>