media: ti: j721e-csi2rx: Fix races while restarting DMA
commit 10993b78334c6d4efb329f5b4d4e284982ebaa74 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit ad79c9ecea5baa7b4f19677e4b1c881ed89b0c3b upstream

After the frame is submitted to DMA, it may happen that the submitted
list is not updated soon enough, and the DMA callback is triggered
before that.

This can lead to kernel crashes, so move everything in a single
lock/unlock section to prevent such races.

Fixes: b4a3d877dc92 ("media: ti: Add CSI2RX support for J721E")
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>