media: ti: j721e-csi2rx: allocate DMA channel based on context index
commit 3740b1bf1c14391acabdb5dad8cc1c6914880220 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

With multiple contexts, there needs to be a different DMA channel for
each context. Earlier, the DMA channel name was hard coded to "rx0" for
the sake of simplicity. Generate the DMA channel name based on its index
and get the channel corresponding to the context.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>