media: i2c: ds90ub960: Support multi-channel sensors
commit f158f8cf7a9c5f673eb0e342df4fffca2341ff69 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Before this, the deserializer would ignore the virtual channel (VC) ID
set in the incoming CSI packets, and set the ID to the RX port number in
the outgoing packets. This was done to support multiple single-channel
cameras, all sending packets with the same default VC ID = 0.

Now we check which (and how many) channels are under use on any incoming
FPDLink port, and map it to available channels on the outgoing CSI port.

Signed-off-by: Jai Luthra <me@jailuthra.in>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>