phy: cadence-torrent: remove noop_ops phy operations
commit 37e15a2810c95a269c19adf6dcbb287f262b3691 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

[ Upstream commit 81472a55edce473cfc22b340c1026c6d5219ccfe ]

Even if a PHY is already configured, the PHY operations are needed during
resume stage, as the PHY is in reset state.
The noop_ops PHY operations is removed to always have PHY operations.
The already_configured flag is checked at the begening of init, configure
and poweron operations to keep the already_configured behaviour.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>
Link: https://lore.kernel.org/r/20240412-j7200-phy-s2r-v1-7-f15815833974@bootlin.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>