counter/ti-eqep: add support for latched position
commit f77887bf5518d45d6806099885cdea4b86b98671 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This adds support to the TI eQEP counter driver for a latched position.
This is a new extension that gets the counter count that was recorded
when an event was triggered. A new device-level latch_mode attribute is
added to select the trigger. Edge capture unit support will be needed
to make full use of this, but "Unit timeout" mode can already be used
to calculate high speeds.

The unit timer could also have attributes for latched_time and
latched_period that use the same trigger. However this is not a use
case at this time, so they can be added later if needed.

Link: https://www.spinics.net/lists/linux-iio/msg64842.html

Signed-off-by: David Lechner <david@lechnology.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>