arm64: dts: ti: k3-j7xx: Change timer nodes status to reserved
commit 448809ccdc0578093bbf625763549c78f2f4823a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The remoteproc firmware like of R5F and DSPs in the MAIN voltage domain
use timers[0]. Therefore, change the status of the timer nodes to
"reserved" to avoid any clash.

This change is already incorporated for timer nodes in the MCU voltage
domain.

[0]: https://software-dl.ti.com/jacinto7/esd/processor-sdk-rtos-jacinto7/09_02_00_05/exports/docs/pdk_jacinto_09_02_00_30/docs/userguide/jacinto/modules/osal.html#rtos-tick-timer-allocation

Fixes: 7f209dd1267c ("arm64: dts: ti: k3-j721e: Add general purpose timers")

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>