arm64: dts: ti: k3-am62a7-sk: Enable ipc with remote proc nodes
commit a82cef91b301e3a03a4efe0f49e6cb8cf50f43af from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Reserve memory for remote rpoc IPC and bind the mailbox assignments
for each remote proc. Two memory regions are reserved for each
remote processor. The first region of 1MB of memory is used for Vring
shared buffers and the second region is used as external memory to the
remote processor, resource table and as tracebuffer.

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>