arm64: dts: ti: k3-j722s-evm: Add support for multiple CAN instances
commit 33665a21aeb826b802f1bc108acc2e27a3f07472 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

CAN instances 0 and 1 in the mcu domain and 0 in the main domain are
brought on the evm through headers J5, J8 and J10 respectively. Thus,
add their respective transceiver's 0, 1 and 2 dt nodes as well as
add the required pinmux to add support for these CAN instances.

Signed-off-by: Bhavya Kapoor <b-kapoor@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>