dt-bindings: net: dp8386x: Add MIT license along with GPL-2.0
commit a2d07f89a32aed544302a4bf56eab6fda29e58fd from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit b472b996a43404a912c5cb4f27050022fdbce10c upstream.

Modify license to include dual licensing as GPL-2.0-only OR MIT
license for TI specific phy header files. This allows for Linux
kernel files to be used in other Operating System ecosystems
such as Zephyr or FreeBSD.

While at this, update the GPL-2.0 to be GPL-2.0-only to be in sync
with latest SPDX conventions (GPL-2.0 is deprecated).

While at this, update the TI copyright year to sync with current year
to indicate license change.

Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Trent Piepho <tpiepho@impinj.com>
Cc: Wadim Egorov <w.egorov@phytec.de>
Cc: Kip Broadhurst <kbroadhurst@ti.com>
Signed-off-by: Udit Kumar <u-kumar1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>