mailbox: add microchip ipc support
commit a48b9dcf9dc7aae7f3152a7a3b230496cda973bc from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

Add a mailbox controller driver for the Microchip Inter-processor
Communication (IPC), which is used to send and receive data between
processors.

This SBI interface in this driver is compatible with the Mi-V Inter-hart
Communication (IHC) IP.

Transmitting and receiving data through the mailbox framework is done
through struct mchp_ipc_msg.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>