dt-bindings: riscv: microchip: document icicle reference design v202406
commit a335609a67a2b80dbabbbb78962f0718fabb1cf4 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

The icicle kit reference design's v2024.06 release includes the
Mi-IHC IP v2, used to send/receive data between clusters when
using Asymmetric Multiprocessing (AMP) mode.

In reference design releases prior to v2024.06, the MI-V IHC subsystem
was included as a proof of concept in the design prior to becoming an
IP available in the Libero catalog.

Among other improvements, the new Mi-V IHC IP v2 includes some
changes to the register map. For this reason, make use of a new
compatible to denote that v2024.06 reference design releases are not
backwards compatible.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>