arm64: dts: ti: k3-am62a-main: SRAM Allocation for VPU
commit fe16b844a800248ff1c5365ac073a70b0f8d9cfa from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The wave5 VPU has a separate AXI interface to on chip SRAM
which can be used for faster access to internal data being
used for processing.

Allow VPU to allocate 62KiB of SRAM memory so that it can
use the same for storing and accessing temporal buffers
while any of the VPU instance is running.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>