octeontx2-pf: Add packet path between representor and VF
commit 5e60d04009d6d561277d4e6dd4f4cc98e78933ad from
git@git.assembla.com:cavium/WindRiver.linux.git

Current HW, do not support in-built switch which will forward pkts
between representee and representor. When representor is put under
a bridge and pkts needs to be sent to representee, then pkts from
representor are sent on a HW internal loopback channel, which again
will be punted to ingress pkt parser. Now the rules that this patch
installs are the MCAM filters/rules which will match against these
pkts and forward them to representee.
The rules that this patch installs are for basic
representor <=> representee path similar to Tun/TAP between VM and
Host.

Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Change-Id: I38b50d18fbf50953bbaf296e75f1ce5651ff10dc
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/129625
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <Bo.Sun.CN@windriver.com>