docs: networking: ti: add driver doc for CPSW Proxy Client
commit 8509c70d4037b8e87a8077fffcfe025476159be9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The CPSW Proxy Client driver interfaces with Ethernet Switch Firmware on
a remote core to enable Ethernet functionality for applications running
on Linux. The Ethernet Switch Firmware (EthFw) is in control of the CPSW
Ethernet Switch on the SoC and acts as the Server, offering services to
Clients running on various cores.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>