net: ethernet: ti: add RPMsg structures for Ethernet Switch Firmware
commit 110d18afc29cd1fa40686dacc940803f6881baf1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The Ethernet Switch Firmware (EthFw) running on the remote MAIN R5 core is
capable of configuring the CPSWnG instance of TI's CPSW Ethernet Switch
on various TI SoCs such as J721E, J7200 and J784S4. EthFw is in complete
control of CPSWnG and acts as a server which allocates and provides the
resources to clients on different cores. All communication between EthFw
and the clients occurs via RPMsg.

Define the message structures used to communicate with EthFw. This shall
enable the Linux Client to avail the services provided by EthFw.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>