net: ethernet: ti: cpsw-proxy-client: enable message exchange with EthFw
commit 152152159ac267da221c373a2cb4109b2fca1035 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a new function which sends RPMsg requests to EthFw and shares the
response for the request. The RPMsg callback function copies the response
it receives from EthFw to the driver's private member, thereby allowing
the response to be shared with the newly added function which sends the
request.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>