net: ethernet: ti: cpsw-proxy-client: add helper to init RX DMA Channels
commit e586a03bf092a330470f778e69e9af664fc952b0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the "init_rx_chans()" function to initialize the RX DMA Channels.
With the knowledge of the PSI-L Thread ID for the RX Channel along with
the details of the RX Flow Base and RX Flow Offset, the RX DMA Flow on
the RX Channel can be setup using the DMA APIs.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>