media: vxe-vxd: decoder: Error handling of fatal condition
commit e0daf9e95f3b4787a15a7f5a1222d84db2dd97f7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Handling of the fatal error response from the IP, setting
the state of frame to discard and resetting the IP. The error
recovery function is executed in the error simulation condition,
when emergency flag is not set, only in a non fatal condition.
When the emergency flag is set, the IP reset is handled in the
schedule_work() api of Interrupt handling.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>