media: platform: img: vxd: fix the error handling
commit 7348e7c98d2efc1d0fcb28edd841f53aa7fd59be from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

fatal errors were not being treated differently to frame errors, so if
there were any decode errors or stream errors the stream was just aborted.

This patch fixes this issue by dividing into error categories, in the case
of fatal errors the stream gets terminated as it previously did.  However
in the case of recoverable errors we just signal back to v4l2 that a frame
was corrupted, but still go ahead and pass the picture data back without
tearing down the stream.  This allows the driver to do any error concealment
on the decoded picture, but still indicate to v4l2 that the frame contains errors.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Acked-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>