net: ethernet: ti: icssg_prueth: introduce and use prueth_swdata struct for SWDATA
commit b4d9c256c6b7cf9160ac3e34279fda1564427751 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

We have different cases for SWDATA (skb, page, cmd, etc)
so it is better to have a dedicated data structure for that.
We can embed the type field inside the struct and use it
to interpret the data in completion handlers.

Increase SWDATA size to 48 so we have some room to add
more data if required.

Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>