spi: spi-mem: Tell controller when device is ready for calibration
commit 43076b8ca30ab9a732edc512dc0c71062c65c326 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Some controllers like the Cadence OSPI controller need to perform a
calibration sequence to operate at high clock speeds. This calibration
should happen after the flash is fully initialized otherwise the
calibration might happen in a different SPI mode from the one the flash
is finally set to. Add a hook that can be used to tell the controller
when the flash is ready for calibration. Whether calibration is needed
depends on the controller.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>