arm64: dts: ti: k3-am62*-sk: Remove HDMI for Fastboot
commit 65c52dce31684b3d37703af6bdf1c5e26993cc67 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In the fastboot scenario where display has to come up quickly
on the OLDI panel, the sii902 bridge driver is not made
available. This leads to the probe deferral of the TIDSS
driver and eventually, failure.

This overlay removes the HDMI support for TI K3 AM62* SoCs

Reviewed-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Aashvij Shenai <a-shenai@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>