arm64: dts: ti: k3-j784s4-evm: Make exp2 gpio-14 usable for both audio and MCAN
commit 9effdef326c302eaac3387d556e439fd7a02e2c8 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

gpio-14 for exp2 is hogged for audio making it busy for mux1 mux-controller
causing can-phy3 probe failure.
So, instead of hogging the gpio as high, mark its idle-state as high in
mux1 node so that it can be used by mux-controller and McASP clock signals
are routed as well irrespective of the mcan being used or not.

Fixes: 100e2c908ea1 ("arm64: dts: ti: k3-j784s4-evm: Enable analog audio support")
Signed-off-by: Jayesh Choudhary <j-choudhary@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>