arm64: dts: ti: k3-am62x-sk-common: Enable wakeup r5f remote proc node
commit cdaa8fa51059a4e008bf75ddfb0f7e86c69994f6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Reserve memory for remote rpoc IPC and bind the mailbox assignments
for each remote proc. Two memory regions are reserved for each
remote processor. The first region of 1Mb of memory is used for Vring
shared buffers and the second region is used as external memory to the
remote processor, resource table and as tracebuffer.

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>