dt-bindings: mfd: syscon: Add ti,am62p-cpsw-mac-efuse compatible
commit 9ca4f77e9808d8d1d0921c36a2459ef0d714d3e2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 3088ab3f190af29c083e71bfd8188d8a92b66abc upstream.

The CTRLMMR_MAC_IDx registers within the CTRL_MMR space of TI's AM62p SoC
contain the MAC Address programmed in the eFuse. Add compatible for
allowing the CPSW driver to obtain a regmap for the CTRLMMR_MAC_IDx
registers within the System Controller device-tree node. The default MAC
Address for the interface corresponding to the first MAC port will be set
to the value programmed in the eFuse.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20240402105708.4114146-1-s-vadapalli@ti.com
Signed-off-by: Lee Jones <lee@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>