media: chips-media: wave5: Add Support for Frequency Points
commit c08bc57d0e3ef0462de5d5c57f58f98cd9d62f4c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Support for OPP table is added to the driver. Driver can now select the
frequency from the available operational frequencies from the OPP table and
run at the frequency that is required for the Pixel rate that the IP is
handling. The Operating frequency adjusts to higher values when multiple
instances are initialized asynchronously and vice-versa. When no
instance is running, the frequency is set to the lowest operational
frequency of the OPP table.

Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>