Revert "drivers: clk: zynqmp: update divider round rate logic"
This patch reverts the commit c249ef9d0978a ("drivers: clk: zynqmp:
update divider round rate logic").

This is because that there has been a similar commit efe0d3640f6ff
("drivers: clk: zynqmp: update divider round rate logic") which is
picked from SDK. This commit introduces some problem and the SDK patch
7e6654b576a00 ("drivers: clk: zynqmp: add hack to use old algorithm
for divider round rate") is to fix it. But when merging, the content of
the commit 7e6654b576a00 is missing. So we need to revert the commit
c249ef9d0978a and bring the content of the commit 7e6654b576a00 back.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>