remoteproc: k3-m4: Move suspend to suspend_late
commit 49b236a4e51e804a1363289c6bdb6d7fb1037e78 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The remote core has to be able to query the DM for the next system mode
in the suspend path. To support this, ti_sci.c has to send the
prepare_sleep command before the suspend message is sent to the remote
core. This patch moves the suspend to suspend_late to be executed after
ti_sci's suspend call.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>