mailbox: omap: Merge mailbox child node setup loops
commit f52c70882b03c4b3f0ddfa0f9e3d2467f7af24be from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Currently the driver loops through all mailbox child nodes twice, once
to read in data from each node, and again to make use of this data.
Instead read the data and make use of it in one pass. This removes
the need for several temporary data structures and reduces the
complexity of this main loop in probe.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Jassi Brar <jassisinghbrar@gmail.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>