mailbox: omap: Use mbox_controller channel list directly
commit c9420a2c4358ddb70b50e7c8514ce5aab442e177 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The driver stores a list of omap_mbox structs so it can later use it to
lookup the mailbox names in of_xlate. This same information is already
available in the mbox_controller passed into of_xlate. Simply use that
data and remove the extra allocation and storage of the omap_mbox list.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Jassi Brar <jassisinghbrar@gmail.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>