mailbox: omap: Remove mbox_chan_to_omap_mbox()
commit 2e28549c34134ebb8d49ab59938270d2997e4519 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This function only checks if mbox_chan *chan is not NULL, but that cannot
be the case and if it was returning NULL which is not later checked
doesn't save us from this. The second check for chan->con_priv is
completely redundant as if it was NULL we would return NULL just the
same. Simply dereference con_priv directly and remove this function.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Jassi Brar <jassisinghbrar@gmail.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>