mailbox: omap: Remove kernel FIFO message queuing
commit b3dd70d5611b758cfff55e2f8d4fc27895d0e075 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The kernel FIFO queue has a couple issues. The biggest issue is that
it causes extra latency in a path that can be used in real-time tasks,
such as communication with real-time remote processors.

The whole FIFO idea itself looks to be a leftover from before the
unified mailbox framework. The current mailbox framework expects
mbox_chan_received_data() to be called with data immediately as it
arrives. Remove the FIFO and pass the messages to the mailbox
framework directly as part of a threaded IRQ handler.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Jassi Brar <jassisinghbrar@gmail.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>