net: ti: icssg_prueth: Enable 10M Link issue quirk for AM64x
commit 6de6e418c80edfbe08f4a5f851c721bd60c0123b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

There is a HW errata for 10M issues which is present on AM65x but not on
AM64x. The workaround for this errata is to enable IEP1. Since both
AM64x and AM65x shares the same ICSSG firmware, firmware is expecting
IEP1 to be enabled for both AM64x and AM65x. As a result 10M Link fails
on AM64x.

Enable the quirk_10m_link_issue so that IEP1 can be enabled for AM64x

Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Reviewed-by: Daolin Qiu <d-qiu@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>