remoteproc: k3-dsp: use local reset for C7xv dsps on AM62A
commit db88712931433e92a5f73cec57c82c5c7ebee593 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The C7xv deep learning engine dsp core in AM62A has a local reset line
from power state controller. In the absence of using the local reset,
it is in an unknown state and when the module is turned on, CPU starts
executing and may go into a bad state. Set local reset to 'true' for
AM62A, so that the remote proc driver properly asserts the local reset
before enabling the module to load the firmware. And it is deasserted
after firmware load is complete to start the core.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>