arm64: dts: imx8-ss-conn: Remove unrelated changes introduced when merging v5.15/standard/base
The merge commit 82bab3e1d163 ("Merge branch 'v5.15/standard/base' into
v5.15/standard/preempt-rt/base") introduced a lot of unrelated changes.
These affected files should have the same content in both the standard
and rt kernels. The changes in this patch is almost the same as the
output of the following command:
  rediff <(git show 82bab3e1d163 arch/arm64/boot/dts/freescale/imx8-ss-conn.dtsi | \
  sed  -e "s#^++\([^-+]\+\|$\)#-\1#" -e "s#^--\([^-+]\+\|$\)#+\1#" -e "s#^  # #" \
  -e "s#^[-+] # #" -e "s#^ [-+]# #" -e "s#^diff --cc#diff --git#" -e "s#^@@@ [0-9,-]\+#@@#" \
  -e "s#@@@#@@#")

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>